python_converter/requirements.txt
pdf2docx==0.5.8
python-docx==1.1.0
reportlab==4.0.7
camelot-py[cv]==0.11.0
openpyxl==3.1.2
weasyprint==60.1
python-pptx==0.6.23
pdfplumber==0.10.3
Pillow==10.1.0
